fixes to match model api

+++ backend/model-behavior/application/src/main/java/banz/ai/marketing/bot/modelbehavior/behavior/entity/ModelResponse.java
@@ -20,7 +20,7 @@ public class ModelResponse {
     private Long id;
 
     @Column(name = "dialog_evaluation")
-    private int dialogEvaluation;
+    private float dialogEvaluation;
 
     @Column(name = "offer_purchase")
     private boolean offerPurchase;
